Added a lot of files
Added files and moved some basic component files around. Getting started on the "demo GUI" that will be part of the presentation. We are focusing on displaying live data (use fakadata.txt). Kelsey is in charge of displaying the live graph, Jacob will display the live data in the text boxes.